Project Description
This is an indicator light to monitor the level of danger in the future. This indicator light is the best way for humans to predict danger in the future. In the future, everyone will be equipped with this indicator light. In the future world, with the development of science and technology, weapons have become more advanced, and there will be frequent violent incidents. Turn on the danger monitoring system. When the system detects that a danger is approaching, the system will automatically analyze the degree of danger, and then light up the indicator light according to the degree of danger, and then activate the corresponding defense system. Note: Green, yellow and orange lights indicate low hazard level, red light indicates high hazard level, and red light is divided into three levels. When the last red light is on, it indicates the highest level of hazard.
Sorry, I want to use a switch to start this indicator system, but I tried countless times, but the switch still doesn't work. I don't know if there is a problem with the circuit or the code, or both? I need help!